feat: implement ProductCard component
