+++ client-app/shared/catalog/components/product-card-grid.vue
@@ -242,6 +242,8 @@ interface IProps {
   productReviewsEnabled?: boolean;
 }
 
+console.warn("ProductCardGrid is deprecated. Use VcProductCard or ProductCard instead.");
+
 const swiperInstance = ref<SwiperInstance>();
 const swiperBulletsState = ref<boolean[]>([true, false, false]);
